+++ CLox/chunk.h
@@ -34,6 +34,6 @@ void initChunk(Chunk* chunk);
 void freeChunk(Chunk* chunk);
 void writeChunk(Chunk* chunk, uint8_t byte, int line);
 bool writeConstant(Chunk* chunk, Value value, int line);
-int getLine(const Chunk* chunk, int instruction);
+int getLine(const Chunk* chunk, size_t instruction);
 
 #endif //clox_chunk_h
